requirements.txt
asgiref
crispy-bootstrap5
Django
django-autoslug
django-cors-headers
django-crispy-forms
django-filter
djangorestframework
Markdown
pillow
pytz
sqlparse
